+++ source/blender/editors/interface/interface_utils.cc
@@ -494,7 +494,7 @@ static bool add_collection_search_item(CollItemSearch *cis,
                             cis->name,
                             cis->data,
                             cis->iconid,
-                            cis->has_sep_char ? (uint64_t)UI_BUT_HAS_SEP_CHAR : 0,
+                            cis->has_sep_char ? (int)UI_BUT_HAS_SEP_CHAR : 0,
                             name_prefix_offset);
 }
